step towards making GodotXUnitApi nuget package: move logic from GodotTestRunner to GodotXUnitRunner, which sits in GodotXUnitApi
