+++ GodotXUnit.csproj
@@ -20,7 +20,7 @@
     <Compile Include="addons\GodotXUnit\Plugin.cs" />
     <Compile Include="addons\GodotXUnit\GodotTestRunner.cs" />
     <Compile Include="addons\GodotXUnit\XUnitDock.cs" />
-    <Compile Include="RiderTestRunner\Runner.cs" />
+    <Compile Include="RiderTestRunner\Runner.cs" Condition="'$(Configuration)' == 'Debug' " />
     <Compile Include="tests\ClickTest.cs" />
     <Compile Include="tests\DebugDrawingTest.cs" />
     <Compile Include="tests\PhysicsCollisionTest.cs" />
@@ -45,7 +45,7 @@
     <Content Include="addons\GodotXUnit\runner\GodotTestRunnerScene.tscn" />
     <Content Include="addons\GodotXUnit\XUnitDock.tscn" />
     <Content Include="addons\GodotXUnit\_work\.gdignore" />
-    <Content Include="RiderTestRunner\Runner.tscn" />
+    <Content Include="RiderTestRunner\Runner.tscn" Condition="'$(Configuration)' == 'Debug' " />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="addons\GodotXUnit\GodotXUnitApi\GodotXUnitApi.csproj">
