+++ RiderTestRunner/Runner.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=2]
 
-[ext_resource path="res://RiderTestRunner/Runner.cs" type="Script" id=1]
+[ext_resource path="Runner.cs" type="Script" id=1]
 
 [node name="Node2D" type="Node2D"]
 script = ExtResource( 1 )
