+++ addons/GodotXUnit/GodotXUnitApi/GodotXUnitApi.csproj
@@ -50,7 +50,7 @@
         <Compile Include="Internal\GodotFactDiscoverer.cs" />
         <Compile Include="Internal\GodotTestCase.cs" />
         <Compile Include="Internal\GodotTestOutputHelper.cs" />
-        <Compile Include="Internal\GodotXUnitRunner.cs" />
+        <Compile Include="Internal\GodotXUnitRunnerBase.cs" />
         <Compile Include="Internal\ProjectListing.cs" />
         <Compile Include="Internal\WorkFiles.cs" />
         <Compile Include="Internal\MessagePassing.cs" />
